Moving extension functions to different package.
