+++ javalin/src/test/java/io/javalin/TestRequest.kt
@@ -9,7 +9,7 @@ package io.javalin
 import com.mashape.unirest.http.Unirest
 import io.javalin.core.security.BasicAuthFilter
 import io.javalin.core.util.Header
-import io.javalin.http.header
+import io.javalin.http.context.header
 import io.javalin.testing.TestUtil
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
